0.4 release
