+++ app/templates/nav.html
@@ -602,6 +602,7 @@
                         {{nav_link(name="Typography", url="/typography.html", active=name)}}
                         {{nav_link(name="Icons", url="/icons.html", active=name)}}
                         {{nav_link(name="Grid", url="/grid.html", active=name)}}
+                        {{nav_link(name="Forms", url="/forms.html", active=name)}}
                     </ul>
                     <!-- /.nav-second-level -->
                 </li>
