+++ app/templates/pages/settings.html
@@ -58,6 +58,19 @@
                         <span class="glyphicon glyphicon-file"></span> zKillboard Watch
                     </div>
                 </div>
+                <div class="row">
+                    <div class="col-md-6">
+                        <div class="form-group">
+                        <input class="form-control" placeholder="typeName"><br>
+                        <button type="button" class="btn btn-danger">Add Item</button>
+                    </div>
+                    </div>
+                    <div class="col-md-6">
+                        <button type="button" class="btn btn-danger">Enable All</button><p></p>
+                        <button type="button" class="btn btn-danger">Disable All</button><p></p>
+                        <button type="button" class="btn btn-danger">Import All</button>
+                    </div>
+                </div>
                 <table data-toggle="table"
                        data-sort-name="logID"
                        data-sort-order="desc"
