+++ consumer_zkillboard.py
@@ -12,17 +12,18 @@
 #-----------------------------------------------------------------------------
 
 from _utility import *
+from _consumer_kills import *
 from requests.exceptions import ConnectionError, ChunkedEncodingError
 import requests.packages.urllib3
 
 requests.packages.urllib3.disable_warnings()
 #  Suppress InsecurePlatformWarning messages
 
-pageNum = 1
 ships = getzkbships()
 r = requests.Response()
 
 for ship in ships:
+    pageNum = 1
     while r.text != "[]":
         start_time = time.time()
         url = 'https://zkillboard.com/api/losses/shipID/' + str(ship[0]) + '/orderDirection/desc/page/' + str(pageNum) + "/"
@@ -55,8 +56,9 @@ for ship in ships:
         detail = "[zkb][typeID:" + str(ship[0]) + "] insert " + str(killmailInsertCount-1) + " @ " + str(round((killmailInsertCount-1)/(time.time() - start_time), 3)) + " rec/sec"
         insertlog_timestamp(service, 0, detail, timestamp)
 
-        print("Completed Page: " + str(pageNum))
+        setzkblastpage(killID, pageNum)  # Keep track of paging
         pageNum += 1
+        print("Completed Page: " + str(pageNum))
 
     setzkbshipenable(ship[0], 0)  # Successful API sets enabled to false
     setzkbshipresult(ship[0], 1)  # Successful API sets importResult to true
